feat: Move Interapp to Intents 👌
